+++ src/PageManagerServiceProvider.php
@@ -4,6 +4,7 @@ namespace Backpack\PageManager;
 
 use Illuminate\Support\ServiceProvider;
 use Illuminate\Routing\Router;
+use Route;
 
 class PageManagerServiceProvider extends ServiceProvider
 {
@@ -38,7 +39,24 @@ class PageManagerServiceProvider extends ServiceProvider
     public function setupRoutes(Router $router)
     {
         $router->group(['namespace' => 'Backpack\PageManager\app\Http\Controllers'], function ($router) {
-            require __DIR__.'/app/Http/routes.php';
+            // Admin Interface Routes
+            Route::group(['middleware' => ['web', 'admin'], 'prefix' => 'admin'], function () {
+                // Backpack\PageManager routes
+                Route::get('page/create/{template}', 'Admin\PageCrudController@create');
+                Route::get('page/{id}/edit/{template}', 'Admin\PageCrudController@edit');
+
+                // This triggered an error before publishing the PageTemplates trait, when calling Route::controller();
+                // CRUD::resource('page', 'Admin\PageCrudController');
+
+                // So for PageCrudController all routes are explicitly defined:
+                Route::get('page/reorder', 'Admin\PageCrudController@reorder');
+                Route::get('page/reorder/{lang}', 'Admin\PageCrudController@reorder');
+                Route::post('page/reorder', 'Admin\PageCrudController@saveReorder');
+                Route::post('page/reorder/{lang}', 'Admin\PageCrudController@saveReorder');
+                Route::get('page/{id}/details', 'Admin\PageCrudController@showDetailsRow');
+                Route::get('page/{id}/translate/{lang}', 'Admin\PageCrudController@translateItem');
+                Route::resource('page', 'Admin\PageCrudController');
+            });
         });
     }